Comments & Output fixes

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Vision/Detectors/NavTargetDetector.java
@@ -187,7 +187,7 @@ public class NavTargetDetector {
     public String positionToString() {
         String output = "(";
         for (double num : getRobotPosition()) {
-            output += num + ", ";
+            output += (int) num + ", ";
         }
 
         return output + ")";
